requirements.txt
python-telegram-bot==11.0.0
PyMySQL=0.9.2